[clean] create ディレクトリを作成しファイルをまとめ, package を変更
どこに何のファイルがあるのか、わかりやすく整理するため
