+++ src/main/java/elasticsearch/sample/create/CreateIndex.java
@@ -1,4 +1,4 @@
-package elasticsearch.sample;
+package elasticsearch.sample.create;
 
 import java.io.IOException;
